requirements/requirements-dev.txt
apscheduler
asgiref
breeze-connect
certifi
chardet
dj-database-url
Django
django-background-tasks
django-heroku
django-sslserver
docopt
et-xmlfile
fuzzymatcher
fuzzywuzzy
gotolong
gunicorn
idna
jdcal
kiteconnect
mysql-connector
mysqlclient
numpy
openpyxl
py5paisa
pandas
plotly
prettytable
psycopg2
pycryptodome
python-dateutil
pytz
pyexcel
pyexcel-xls
requests
six
smartapi-python
sqlparse
xlrd
urllib3
websocket-client
wcwidth
whitenoise
yarg
